Add blog installer class
Move the installer code to a separate class.
